+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 -   Added bookmarks and a bookmarks system.
 -   Added ability to set all settings to defaults.
 -   Added google analytics to measure if people use it or not. Feel free to block this script, and let me know if you want it to be removed.
+-   Tweaked weather CSS.